adding support for Docker as a new container technology
