+++ .github/workflows/test.yml
@@ -10,7 +10,7 @@ jobs:
       fail-fast: false
       matrix:
         module: ["lmod", "tcl"]
-        container_tech: ["podman", "singularity"]
+        container_tech: ["podman", "singularity", "docker"]
     steps:
 
       - name: Install Dependencies
@@ -62,6 +62,7 @@ jobs:
           source activate shpc
           shpc config set container_tech:${{ matrix.container_tech }}
           shpc config set module_sys:${{ matrix.module }}
+          shpc config set enable_tty:false
           shpc install python:3.9.5-alpine
           module use ./modules
           module load python/3.9.5-alpine
@@ -87,6 +88,7 @@ jobs:
           source activate shpc
           shpc config set container_tech:${{ matrix.container_tech }}
           shpc config set module_sys:${{ matrix.module }}
+          shpc config set enable_tty:false
           shpc install python:3.9.5-alpine
 
           shopt expand_aliases || true
