+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.scom/singularityhub/singularity-hpc/tree/master) (0.0.x)
+ - Docker support (0.0.26)
+   - added an enable_tty setting to allow disabling add -t in recipes
+   - enforced not adding extra settings to setting.yml
+   - settings are validated when an update is attempted
  - Podman support (0.0.25)
    - container_tech is now a settings.yml variable
    - allow for custom test interpreter "test_shell"
